bug fix
Fixed login and register not working on android ( issue is with useRef ), app works normal on web but on android env. the ref thows undefined value, Had to use useState to get the values ( downside is that the component re-renders on every key stroke). Started implementing Tools. App works perfect on web and android.